gca-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
torch==2.5.1
google-genai==1.0.0
python-multipart==0.0.17
pyyaml==6.0.2
